# specs/001-proposal-assistant/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: EasyGrant Smart Proposal Assistant API
  description: |
    REST API for the Smart Proposal Assistant demo. Supports funding call upload,
    requirements extraction, RAG-based section generation with citations, sticky edits,
    and DOCX export.
  version: 1.0.0
  contact:
    name: EasyGrant Team
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://easygrant.onrender.com
    description: Production (Render)
...
paths:
...
  /api/sections/generate:
    post:
      tags: [Sections]
      summary: Generate a section draft
      description: |
        Generate draft text for a specific section using RAG retrieval from
        uploaded documents. Includes inline citations and enforces word limits.
        (FR-005, FR-006, FR-007)
      operationId: generateSection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateSectionRequest'
      responses:
        '200':
          description: Section generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedSectionResponse'
        '400':
          description: Invalid request (missing section_id, no supporting docs)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/sections/{section_id}:
    get:
      tags: [Sections]
      summary: Get section details
      description: Retrieve generated section with current text, citations, and metadata
      operationId: getSection
      parameters:
        - name: section_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Section details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedSectionResponse'
        '404':
          description: Section not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    patch:
      tags: [Sections]
      summary: Update section (user edits)
      description: |
        Apply user edits to a section. Edits are stored separately to enable
        sticky edit preservation during regeneration. (FR-009)
      operationId: updateSection
      parameters:
        - name: section_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSectionRequest'
      responses:
        '200':
          description: Section updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedSectionResponse'
...
components:
  schemas:
...
    GenerateSectionRequest:
      type: object
      properties:
        session_id:
          type: string
          format: uuid
        section_id:
          type: string
          format: uuid
          description: ID of SectionRequirement to generate
        regenerate:
          type: boolean
          default: false
          description: If true, regenerate existing section
        feedback:
          type: string
          nullable: true
          description: User feedback to guide generation
      required: [session_id, section_id]
    GeneratedSectionResponse:
      type: object
      properties:
        section_id:
          type: string
          format: uuid
        section_name:
          type: string
          example: "Community Need Statement"
        final_text:
          type: string
          description: Merged AI text + user edits
        word_count:
          type: integer
          example: 487
        word_limit:
          type: integer
          nullable: true
          example: 500
        limit_exceeded:
          type: boolean
          example: false
        citations:
          type: array
          items:
            $ref: '#/components/schemas/Citation'
        is_locked:
          type: boolean
          example: false
        generated_at:
          type: string
          format: date-time
        regeneration_count:
          type: integer
          example: 0
    Citation:
      type: object
      properties:
        citation_id:
          type: string
        position_in_text:
          type: integer
          description: Character offset in final_text
        document_id:
          type: string
          format: uuid
        document_title:
          type: string
          example: "Strategic Plan 2024"
        page_number:
          type: integer
          example: 5
        chunk_text:
          type: string
          description: Excerpt from source chunk
          example: "...our community lacks reliable internet..."
        relevance_score:
          type: number
          format: float
          example: 0.87
    UpdateSectionRequest:
      type: object
      properties:
        edited_text:
          type: string
          description: New text with user edits applied
      required: [edited_text]
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "File size exceeds 10MB limit"
        code:
          type: string
          example: "FILE_TOO_LARGE"
        details:
          type: object
          nullable: true